refactor(engine): split markdown.ts into 8 focused modules
Splits the 1,070-line markdown processing monolith into:
- heading-id.ts (shared leaf — breaks circular dep)
- types.ts (all interfaces: Header, Post, Page, ContentLoader, etc.)
- renderer.ts (markdown-it config, plugins, renderMarkdown)
- parse.ts (extractHeaders, parseMarkdownContent, processAnchorTags)
- gutter.ts (processGutterContent)
- modules.ts (processMarkdownModules, getItemBySlug, getPageByFilename)
- content-loader.ts (createContentLoader factory)
- registry.ts (global content loader registry)

Original markdown.ts becomes a thin re-export barrel — all existing
import paths continue to work unchanged.
