+++ libs/engine/src/lib/content/markdown/heading-id.ts
@@ -1,0 +1,8 @@
+export function generateHeadingId(text: string): string {
+	return text
+		.toLowerCase()
+		.replace(/[^\w\s-]/g, "")
+		.replace(/\s+/g, "-")
+		.replace(/-+/g, "-")
+		.trim();
+}
